Move Object-related types + Rename one type file
