+++ algoliasearch/types_delete.go
@@ -1,0 +1,5 @@
+package algoliasearch
+
+type DeleteRes struct {
+	DeletedAt string `json:"deletedAt"`
+}
