+++ algoliasearch/types_object.go
@@ -2,6 +2,18 @@ package algoliasearch
 
 import "fmt"
 
+type CreateObjectRes struct {
+	CreatedAt string `json:"createdAt"`
+	ObjectID  string `json:"objectID"`
+	TaskID    int    `json:"taskID"`
+}
+
+type UpdateObjectRes struct {
+	ObjectID  string `json:"objectID"`
+	TaskID    int    `json:"taskID"`
+	UpdatedAt string `json:"updatedAt"`
+}
+
 type objects struct {
 	Results []Object `json:"results"`
 }
